pyproject.toml
[project]
name = "doc2md"
version = "0.1.0"
description = "Document to Markdown converter API - PDF, Word, HTML to Markdown"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-multipart>=0.0.17",
    "httpx>=0.27.0",
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "markdownify>=0.13.0",
    "beautifulsoup4>=4.12.0",
    "mcp>=1.0.0",
]

[project.scripts]
doc2md-server = "doc2md.api:main"
doc2md-mcp = "doc2md.mcp_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
